feat: extract PostgreSQL pool to module
