--- models/Request.js
@@ -1,96 +1,0 @@
-const mongoose = require('mongoose');
-
-const RequestSchema = new mongoose.Schema({
-  title: {
-    type: String,
-    required: true,
-    trim: true
-  },
-  description: {
-    type: String,
-    required: true
-  },
-  category: {
-    type: String,
-    required: true,
-    enum: [
-      'medical',
-      'humanitarian', 
-      'evacuation',
-      'psychological',
-      'legal',
-      'technical',
-      'translation',
-      'other'
-    ]
-  },
-  priority: {
-    type: String,
-    enum: ['low', 'medium', 'high', 'urgent'],
-    default: 'medium'
-  },
-  location: {
-    address: {
-      type: String,
-      required: true
-    },
-    city: {
-      type: String,
-      required: true
-    },
-    region: String,
-    coordinates: {
-      lat: Number,
-      lng: Number
-    }
-  },
-  contactInfo: {
-    name: {
-      type: String,
-      required: true
-    },
-    phone: {
-      type: String,
-      required: true
-    },
-    email: String,
-    alternateContact: String
-  },
-  status: {
-    type: String,
-    enum: ['new', 'assigned', 'in_progress', 'completed', 'cancelled'],
-    default: 'new'
-  },
-  assignedVolunteer: {
-    type: mongoose.Schema.Types.ObjectId,
-    ref: 'User',
-    default: null
-  },
-  deadline: Date,
-  notes: [{
-    text: String,
-    author: {
-      type: mongoose.Schema.Types.ObjectId,
-      ref: 'User'
-    },
-    createdAt: {
-      type: Date,
-      default: Date.now
-    }
-  }],
-  createdAt: {
-    type: Date,
-    default: Date.now
-  },
-  updatedAt: {
-    type: Date,
-    default: Date.now
-  }
-});
-
-RequestSchema.pre('save', function(next) {
-  this.updatedAt = Date.now();
-  next();
-});
-
-module.exports = mongoose.model('Request', RequestSchema);
